requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5

# Database
psycopg[binary]>=3.2.0

# Storage
boto3==1.34.22
google-cloud-storage==2.14.0

# Environment & Config
python-dotenv==1.0.0

# Authentication
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.0

# Utilities
python-dateutil==2.8.2
